TP2 and  3/services/epf-flower-data-science/requirements.txt
gunicorn~=20.1
uvicorn==0.17.6
fastapi==0.95.1
fastapi-utils==0.2.1
pydantic==1.10
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0
httpx==0.24.1
python-dotenv==1.0.0
kaggle==1.5.13
opendatasets==0.1.22
google-cloud-firestore==2.11.0
pytest==7.4.0
